config: move config to own folder
